requirements.txt
flask>=3.0.0
crewai==0.86.0
crewai-tools>=0.17.0
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-openai>=0.2.0
python-dotenv>=1.0.0
pydantic>=2.9.0
pandas>=2.2.0
numpy>=2.0.0
setuptools>=68.0.0
